Create Util Class for getAccessToken, Add second step (prcessing access token) for API Connection
