+++ backend/src/main/java/de/propertyfinder/backend/controller/TestController.java
@@ -1,6 +1,6 @@
 package de.propertyfinder.backend.controller;
 
-import de.propertyfinder.backend.api.OctoparseApiService;
+import de.propertyfinder.backend.api.octoparse.OctoparseApiService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -16,6 +16,6 @@ public class TestController {
 
     @GetMapping
     public String getData(){
-        return octoparseApiService.testVerifyApi();
+        return octoparseApiService.getAllProperties();
     }
 }
